% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_area.r
\name{bb_area}
\alias{bb_area}
\title{Add a polygonal area}
\usage{
bb_area(
  bb,
  x,
  y,
  fill = "#8888ff",
  alpha = 0.3,
  stroke = "none",
  style = nlist(fill = fill, `fill-opacity` = alpha, stroke = stroke, ...),
  level = -10,
  ...,
  id = random.string(),
  tooltip = NULL
)
}
\arguments{
\item{bb}{A `bb` graphic object.}

\item{x, y}{Vectors containing the polygon coordinates.}

\item{fill}{Fill color.}

\item{alpha}{Fill opacity.}

\item{stroke}{Stroke color.}

\item{style}{A list of SVG style properties.}

\item{level}{Drawing level; lower values are drawn first.}

\item{...}{Additional SVG style properties.}

\item{id}{A unique object identifier.}

\item{tooltip}{Optional tooltip text.}
}
\value{
The modified `bb` object.
}
\description{
Add a polygonal area
}
\examples{
\dontrun{
bb_pane(xrange = c(0, 4), yrange = c(0, 4)) |>
  bb_area(x = c(1, 1, 3), y = c(1, 3, 1))
}
}
