% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_area.r
\name{bb_area_rect}
\alias{bb_area_rect}
\title{Add a rectangular area}
\usage{
bb_area_rect(bb, x1, y1, x2, y2, ...)
}
\arguments{
\item{bb}{A `bb` graphic object.}

\item{x1, y1}{Coordinates of one corner.}

\item{x2, y2}{Coordinates of the opposite corner.}

\item{...}{Additional arguments passed to [bb_area()].}
}
\value{
The modified `bb` object.
}
\description{
Add a rectangular area
}
\examples{
\dontrun{
bb_pane(xrange = c(0, 4), yrange = c(0, 4)) |>
  bb_area_rect(1, 1, 3, 2)
}
}
