% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_line.r
\name{bb_arrow}
\alias{bb_arrow}
\title{Add an arrow}
\usage{
bb_arrow(
  bb,
  x1 = x,
  x2 = x1,
  y1 = y,
  y2 = y1,
  x,
  y,
  arrow.head = c("end"),
  alpha = NULL,
  color = NULL,
  class = "arrow",
  style = list(stroke = color, `stroke-opacity` = alpha, ...),
  ...,
  id = random.string()
)
}
\arguments{
\item{bb}{A `bb` graphic object.}

\item{x1, y1}{Coordinates of the first endpoint.}

\item{x2, y2}{Coordinates of the second endpoint.}

\item{x, y}{Convenience coordinates used by the endpoint defaults.}

\item{arrow.head}{Which endpoint receives an arrow head.}

\item{alpha}{Stroke opacity.}

\item{color}{Stroke color.}

\item{class}{SVG class name.}

\item{style}{A list of SVG style properties.}

\item{...}{Additional SVG style properties.}

\item{id}{A unique object identifier.}
}
\value{
The modified `bb` object.
}
\description{
Add an arrow
}
\examples{
\dontrun{
bb_pane(xrange = c(0, 4), yrange = c(0, 4)) |>
  bb_arrow(x1 = 1, y1 = 1, x2 = 3, y2 = 3)
}
}
