% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_comp.r
\name{bb_compute_obj}
\alias{bb_compute_obj}
\title{Compute one object in a graphic}
\usage{
bb_compute_obj(bb, obj, i)
}
\arguments{
\item{bb}{A `bb` graphic object.}

\item{obj}{An object specification from `bb$objs`.}

\item{i}{The object's position in `bb$objs`.}
}
\value{
The `bb` object with the selected geometry computed.
}
\description{
Compute one object in a graphic
}
\examples{
\dontrun{
bb <- bb_pane(xrange = c(0, 1), yrange = c(0, 1)) |>
  bb_point(0.5, 0.5)
bb_compute_obj(bb, bb$objs[[1]], 1)
}
}
