% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_comp.r
\name{bb_compute_objs}
\alias{bb_compute_objs}
\title{Compute all objects in a graphic}
\usage{
bb_compute_objs(bb)
}
\arguments{
\item{bb}{A `bb` graphic object.}
}
\value{
The `bb` object with computed geometry.
}
\description{
Compute all objects in a graphic
}
\examples{
\dontrun{
bb <- bb_pane(xrange = c(0, 1), yrange = c(0, 1)) |>
  bb_point(0.5, 0.5)
bb_compute_objs(bb)
}
}
