% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_curve.r
\name{bb_compute_slopecurve}
\alias{bb_compute_slopecurve}
\title{Compute a slope-defined curve's coordinates}
\usage{
bb_compute_slopecurve(bb, obj, ...)
}
\arguments{
\item{bb}{A `bb` graphic object.}

\item{obj}{A slope-curve specification.}

\item{...}{Reserved for additional computation options.}
}
\value{
The curve specification with computed geometry.
}
\description{
Compute a slope-defined curve's coordinates
}
\examples{
\dontrun{
bb <- bb_pane(xrange = c(0, 5), yrange = c(0, 5)) |>
  bb_slopecurve(x = 2, y = 2, slope = "-y_ / x_")
bb_compute_slopecurve(bb, bb$objs[[1]])
}
}
