% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb.r
\name{bb_data}
\alias{bb_data}
\title{Attach data to a graphic}
\usage{
bb_data(bb, data)
}
\arguments{
\item{bb}{A `bb` graphic object.}

\item{data}{A data frame or similar object.}
}
\value{
The modified `bb` object.
}
\description{
Attach data to a graphic
}
\examples{
\dontrun{
bb_pane() |> bb_data(data.frame(x = 1:3, y = 3:1))
}
}
