% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb.r
\name{bb_hide_object}
\alias{bb_hide_object}
\title{Hide a graphic object}
\usage{
bb_hide_object(bb, id)
}
\arguments{
\item{bb}{A `bb` graphic object.}

\item{id}{Identifier of the object to hide.}
}
\value{
The modified `bb` object.
}
\description{
Hide a graphic object
}
\examples{
\dontrun{
bb <- bb_pane(xrange = c(0, 5), yrange = c(0, 5)) |>
  bb_point(2, 3, id = "p")
bb_hide_object(bb, "p")
}
}
