% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_curve.r
\name{bb_isoquant}
\alias{bb_isoquant}
\title{Add an isoquant through a point}
\usage{
bb_isoquant(
  bb,
  Q,
  x,
  y,
  ...,
  id = paste0("isoquant_", random.string()),
  xvar = bb$xvar,
  yvar = bb$yvar
)
}
\arguments{
\item{bb}{A `bb` graphic object.}

\item{Q}{A production function or expression accepted by
`isoquant.slope()`.}

\item{x, y}{Coordinates of a point on the isoquant.}

\item{...}{Additional arguments passed to [bb_slopecurve()].}

\item{id}{A unique object identifier.}

\item{xvar, yvar}{Names of the horizontal and vertical variables.}
}
\value{
The modified `bb` object.
}
\description{
Add an isoquant through a point
}
\examples{
\dontrun{
bb_pane(xrange = c(0, 5), yrange = c(0, 5)) |>
  bb_isoquant(Q = "x_ * y_", x = 2, y = 2)
}
}
