% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb.r
\name{bb_margins}
\alias{bb_margins}
\title{Specify margins in pixels for bb object}
\usage{
bb_margins(bb, bottom = NULL, left = NULL, top = NULL, right = NULL, ...)
}
\description{
Specify margins in pixels for bb object
}
