% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb.r
\name{bb_object}
\alias{bb_object}
\title{Add an object specification}
\usage{
bb_object(
  bb,
  obj = NULL,
  ...,
  id = first.non.null(obj[["id"]], random.string())
)
}
\arguments{
\item{bb}{A `bb` graphic object.}

\item{obj}{Optional list describing the object.}

\item{...}{Named properties added to `obj`.}

\item{id}{A unique object identifier.}
}
\value{
The modified `bb` object.
}
\description{
Add an object specification
}
\examples{
\dontrun{
bb <- bb_pane(xrange = c(0, 5), yrange = c(0, 5))
bb_object(bb, list(type = "point", x = 2, y = 3), id = "p")
}
}
