% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb.r
\name{bb_point}
\alias{bb_point}
\title{Draw a point}
\usage{
bb_point(
  bb,
  x,
  y,
  r = 4,
  alpha = NULL,
  color = fill,
  fill = NULL,
  class = "point",
  style = list(stroke = color, `fill-color` = fill, `stroke-opacity` = alpha,
    `fill-opacity` = alpha, ...),
  ...,
  id = paste0("point_", random.string())
)
}
\description{
Draw a point
}
