% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_line.r
\name{bb_segment}
\alias{bb_segment}
\title{Add a line segment}
\usage{
bb_segment(
  bb,
  x1 = x,
  x2 = x1,
  y1 = y,
  y2 = y1,
  x,
  y,
  alpha = NULL,
  color = NULL,
  class = "segment",
  linetype = "solid",
  lwd = NULL,
  dasharray = linetype.to.dasharry(linetype),
  style = list(stroke = color, `stroke-opacity` = alpha, `stroke-width` = lwd, ...),
  ...,
  tooltip = NULL,
  id = paste0("segment_", random.string())
)
}
\arguments{
\item{bb}{A `bb` graphic object.}

\item{x1, y1}{Coordinates of the first endpoint.}

\item{x2, y2}{Coordinates of the second endpoint.}

\item{x, y}{Convenience coordinates used by the endpoint defaults.}

\item{alpha}{Stroke opacity.}

\item{color}{Stroke color.}

\item{class}{SVG class name.}

\item{linetype}{Line type.}

\item{lwd}{Stroke width.}

\item{dasharray}{SVG stroke-dasharray value.}

\item{style}{A list of SVG style properties.}

\item{...}{Additional SVG style properties.}

\item{tooltip}{Optional tooltip text.}

\item{id}{A unique object identifier.}
}
\value{
The modified `bb` object.
}
\description{
Add a line segment
}
\examples{
\dontrun{
bb_pane(xrange = c(0, 4), yrange = c(0, 4)) |>
  bb_segment(x1 = 1, y1 = 1, x2 = 3, y2 = 3)
}
}
