% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_series.r
\name{bb_series_tooltip_bars}
\alias{bb_series_tooltip_bars}
\title{Add tooltip bars for data series}
\usage{
bb_series_tooltip_bars(
  bb,
  xname = "t",
  color = "yellow",
  lwd = 11,
  style = list(stroke = color, `stroke-width` = lwd),
  id = paste0("series_tooltip_bars", random.string()),
  level = 11,
  round.digits = 2,
  signif.digits = 5,
  tooltip.fun = NULL,
  tooltip.data = NULL,
  ...
)
}
\arguments{
\item{bb}{A `bb` graphic object.}

\item{xname}{Label used for the horizontal coordinate in tooltips.}

\item{color}{Bar color.}

\item{lwd}{Bar width.}

\item{style}{A list of SVG style properties.}

\item{id}{A unique object identifier.}

\item{level}{Drawing level.}

\item{round.digits, signif.digits}{Numeric formatting controls.}

\item{tooltip.fun}{Optional function that creates tooltip text.}

\item{tooltip.data}{Optional data passed to `tooltip.fun`.}

\item{...}{Additional object properties.}
}
\value{
The modified `bb` object.
}
\description{
Add tooltip bars for data series
}
\examples{
\dontrun{
bb_pane() |>
  bb_series(x = 1:5, y = 1:5) |>
  bb_series_tooltip_bars()
}
}
