% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_curve.r
\name{bb_slopecurve}
\alias{bb_slopecurve}
\title{Add a curve defined by its slope}
\usage{
bb_slopecurve(
  bb,
  x,
  y,
  slope,
  color = NULL,
  lwd = NULL,
  alpha = NULL,
  style = nlist(stroke = color, stroke_width = lwd, `stroke-opacity` = alpha, ...),
  x.move = 0,
  y.move = 0,
  xrange = bb$xrange,
  yrange = bb$yrange,
  var.funs = NULL,
  tooltip = NULL,
  ...,
  data = NULL,
  id = paste0("slopecurve_", random.string()),
  xvar = bb$xvar,
  yvar = bb$yvar
)
}
\arguments{
\item{bb}{A `bb` graphic object.}

\item{x, y}{Coordinates of a point on the curve.}

\item{slope}{A slope expression or character string.}

\item{color}{Stroke color.}

\item{lwd}{Stroke width.}

\item{alpha}{Stroke opacity.}

\item{style}{A list of SVG style properties.}

\item{x.move, y.move}{Offsets applied to the computed coordinates.}

\item{xrange, yrange}{Two-element coordinate ranges.}

\item{var.funs}{Optional variable functions used in computations.}

\item{tooltip}{Optional tooltip text.}

\item{...}{Additional SVG style properties.}

\item{data}{Optional data used to evaluate expressions.}

\item{id}{A unique object identifier.}

\item{xvar, yvar}{Names of the horizontal and vertical variables.}
}
\value{
The modified `bb` object.
}
\description{
Numerically traces a curve through a supplied point using a slope
expression.
}
\examples{
\dontrun{
bb_pane(xrange = c(0, 5), yrange = c(0, 5)) |>
  bb_slopecurve(x = 2, y = 2, slope = "-y_ / x_")
}
}
