% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb.r
\name{bb_svg_css}
\alias{bb_svg_css}
\title{Default CSS for blackboard-style SVG graphics}
\usage{
bb_svg_css()
}
\value{
A character string containing CSS rules.
}
\description{
Default CSS for blackboard-style SVG graphics
}
\examples{
\dontrun{
css <- bb_svg_css()
}
}
