% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_line.r
\name{bb_tangent}
\alias{bb_tangent}
\title{Add a tangent line}
\usage{
bb_tangent(
  bb,
  x = NULL,
  y = NULL,
  slope = NULL,
  width = NULL,
  alpha = NULL,
  color = NULL,
  class = "segment",
  linetype = "solid",
  to = NULL,
  ...,
  id = paste0("tangent_", random.string())
)
}
\arguments{
\item{bb}{A `bb` graphic object.}

\item{x, y}{Coordinates of the tangency point.}

\item{slope}{The tangent slope.}

\item{width}{Total line width, or a two-element vector of widths on either
side of the tangency point.}

\item{alpha}{Stroke opacity.}

\item{color}{Stroke color.}

\item{class}{SVG class name.}

\item{linetype}{Line type.}

\item{to}{Optional identifier of a curve from which missing tangent
quantities are computed.}

\item{...}{Additional arguments passed to [bb_segment()].}

\item{id}{A unique object identifier.}
}
\value{
The modified `bb` object.
}
\description{
Add a tangent line
}
\examples{
\dontrun{
bb_pane(xrange = c(0, 4), yrange = c(0, 4)) |>
  bb_tangent(x = 2, y = 2, slope = 1)
}
}
