% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_export.r
\name{bb_to_ps}
\alias{bb_to_ps}
\title{Export a graphic to PostScript}
\usage{
bb_to_ps(bb, file, ...)
}
\arguments{
\item{bb}{A `bb` graphic object.}

\item{file}{Output file path.}

\item{...}{Additional arguments passed through the export pipeline.}
}
\value{
The result returned by the SVG conversion function, invisibly where
  applicable.
}
\description{
Export a graphic to PostScript
}
\examples{
\dontrun{
bb <- bb_pane(xrange = c(0, 1), yrange = c(0, 1))
bb_to_ps(bb, "figure.ps")
}
}
