% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb.r
\name{bb_var}
\alias{bb_var}
\title{Define computed graphic variables}
\usage{
bb_var(bb, ..., id = paste0("var_", random.string()))
}
\arguments{
\item{bb}{A `bb` graphic object.}

\item{...}{Named expressions defining variables.}

\item{id}{A unique object identifier.}
}
\value{
The modified `bb` object.
}
\description{
Define computed graphic variables
}
\examples{
\dontrun{
bb_pane(xrange = c(0, 5), yrange = c(0, 5)) |> bb_var(a = 2)
}
}
