% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_line.r
\name{bb_vline}
\alias{bb_vline}
\title{Add a vertical line}
\usage{
bb_vline(
  bb,
  x,
  alpha = NULL,
  color = NULL,
  class = "segment",
  linetype = "solid",
  lwd = NULL,
  style = list(stroke = color, `stroke-opacity` = alpha, `stroke-width` = lwd, ...),
  ...,
  id = paste0("hline_", random.string())
)
}
\arguments{
\item{bb}{A `bb` graphic object.}

\item{x}{Horizontal coordinate.}

\item{alpha}{Stroke opacity.}

\item{color}{Stroke color.}

\item{class}{SVG class name.}

\item{linetype}{Line type.}

\item{lwd}{Stroke width.}

\item{style}{A list of SVG style properties.}

\item{...}{Additional SVG style properties.}

\item{id}{A unique object identifier.}
}
\value{
The modified `bb` object.
}
\description{
Add a vertical line
}
\examples{
\dontrun{
bb_pane(xrange = c(0, 4), yrange = c(0, 4)) |> bb_vline(x = 2)
}
}
