% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb.r
\name{bb_xaxis}
\alias{bb_xaxis}
\title{Specify the x-axis}
\usage{
bb_xaxis(
  bb,
  label = latex,
  latex = NULL,
  labelpos = c("bottom", "right", "center")[1],
  show.ticks = first.non.null(defaults$show.ticks, TRUE),
  arrow.axis = first.non.null(defaults$arrow.axis, !isTRUE(show.ticks)),
  defaults = bb$defaults,
  y.offset = NULL,
  x.offset = NULL,
  y = "bottom",
  align = NULL,
  num.ticks = 5,
  ticks = NULL
)
}
\arguments{
\item{y}{position of the axis. Either "bottom" (default) or "top" or a number}
}
\description{
Specify the x-axis
}
