% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_label.r
\name{bb_xtick}
\alias{bb_xtick}
\title{Add a custom x-axis tick}
\usage{
bb_xtick(
  bb,
  x = NULL,
  ...,
  label = x,
  latex = NULL,
  align = "center",
  y.offset = -20,
  y = NULL,
  id = random.string(),
  add.tick.line = TRUE
)
}
\arguments{
\item{bb}{A `bb` graphic object.}

\item{x}{Horizontal tick coordinate.}

\item{...}{Additional arguments passed to [bb_text()].}

\item{label}{Plain-text tick label.}

\item{latex}{LaTeX tick label.}

\item{align}{Horizontal label alignment.}

\item{y.offset}{Vertical pixel offset.}

\item{y}{Vertical coordinate of the label.}

\item{id}{A unique object identifier.}

\item{add.tick.line}{Whether to add a tick mark to the axis.}
}
\value{
The modified `bb` object.
}
\description{
Add a custom x-axis tick
}
\examples{
\dontrun{
bb_pane(xrange = c(0, 4), yrange = c(0, 4)) |>
  bb_xtick(x = 2, label = "two")
}
}
