% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb.r
\name{bb_ymarker}
\alias{bb_ymarker}
\title{Add a horizontal marker}
\usage{
bb_ymarker(
  bb,
  y = NULL,
  x2 = x,
  x = NULL,
  ...,
  linetype = "dashed",
  label = y,
  latex = NULL,
  align = "right",
  id = random.string()
)
}
\arguments{
\item{bb}{A `bb` graphic object.}

\item{y}{Vertical marker coordinate.}

\item{x2}{Ending horizontal coordinate.}

\item{x}{Optional shorthand for `x2`.}

\item{...}{Additional arguments passed to the tick and segment.}

\item{linetype}{Marker-line type.}

\item{label}{Plain-text marker label.}

\item{latex}{Optional LaTeX marker label.}

\item{align}{Label alignment.}

\item{id}{A base identifier for the marker objects.}
}
\value{
The modified `bb` object.
}
\description{
Add a horizontal marker
}
\examples{
\dontrun{
bb_pane(xrange = c(0, 5), yrange = c(0, 5)) |> bb_ymarker(y = 2)
}
}
