% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb_label.r
\name{bb_ytick}
\alias{bb_ytick}
\title{Add a custom y-axis tick}
\usage{
bb_ytick(
  bb,
  y = NULL,
  ...,
  label = y,
  latex = NULL,
  align = "right",
  x.offset = ifelse(add.tick.line, -15, -5),
  x = NULL,
  id = random.string(),
  add.tick.line = TRUE
)
}
\arguments{
\item{bb}{A `bb` graphic object.}

\item{y}{Vertical tick coordinate.}

\item{...}{Additional arguments passed to [bb_text()].}

\item{label}{Plain-text tick label.}

\item{latex}{LaTeX tick label.}

\item{align}{Horizontal label alignment.}

\item{x.offset}{Horizontal pixel offset.}

\item{x}{Horizontal coordinate of the label.}

\item{id}{A unique object identifier.}

\item{add.tick.line}{Whether to add a tick mark to the axis.}
}
\value{
The modified `bb` object.
}
\description{
Add a custom y-axis tick
}
\examples{
\dontrun{
bb_pane(xrange = c(0, 4), yrange = c(0, 4)) |>
  bb_ytick(y = 2, label = "two")
}
}
