% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb.r
\name{bbsvg}
\alias{bbsvg}
\title{Initiate a bb graphic object}
\usage{
bbsvg(
  bb = NULL,
  id = NULL,
  data = NULL,
  xvar = xy[1],
  yvar = xy[2],
  xy = c("x_", "y_"),
  xrange = NULL,
  yrange = NULL,
  show.ticks = FALSE,
  arrow.axis = NULL,
  xlen = 201,
  ylen = 201,
  org.width = width,
  org.height = height,
  margins = NULL,
  show = ".all",
  hide = NULL,
  init.data = FALSE,
  dataenv = parent.frame(),
  css = bb_svg_css(),
  values = if (!is.null(data)) as.list(data[data.row, , drop = FALSE]) else list(),
  data.row = 1,
  enclos = parent.frame(),
  scale = 1,
  width = 420,
  height = 300,
  ...
)
}
\arguments{
\item{bb}{possibily modify an existing bb object}
}
\description{
Initiate a bb graphic object
}
