% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines_utils.r
\name{geom.curves.intersections}
\alias{geom.curves.intersections}
\title{Find the intersections of two curves, which are characterized by their xy values}
\usage{
geom.curves.intersections(geom1, geom2, grid.length = 201)
}
\description{
Find the intersections of two curves, which are characterized by their xy values
}
