% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positions.r
\name{geom.to.geom.pos}
\alias{geom.to.geom.pos}
\title{Find relative position (above, below, left, right) of a geom to a geom}
\usage{
geom.to.geom.pos(new, old, check = c("above", "below", "left", "right"))
}
\description{
If the geoms intersect non of the attributes holds true, i.e.
the geom new must be stricly above old
}
