% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb.r
\name{bb_set_data}
\alias{bb_set_data}
\title{Set the current data set}
\usage{
bb_set_data(
  values = if (!is.null(data)) as.list(data[data.row, , drop = FALSE]) else list(),
  data = bb$data,
  data.row = first.non.null(bb$data.row, 1)
)
}
\description{
Set the current data set
}
