% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb.r
\name{bb_yaxis}
\alias{bb_yaxis}
\title{Specify the y-axis}
\usage{
bb_yaxis(
  bb,
  label = latex,
  latex = NULL,
  labelpos = c("left", "top", "center")[1],
  show.ticks = first.non.null(defaults$show.ticks, TRUE),
  arrow.axis = first.non.null(defaults$arrow.axis, !isTRUE(show.ticks)),
  defaults = bb$defaults,
  y.offset = NULL,
  x.offset = NULL,
  align = NULL,
  x = "left",
  ticks = NULL,
  num.ticks = 5,
  show.grid = FALSE,
  grid.ticks = ticks,
  grid.color = "#888888",
  tick.labels = NULL,
  show.line = TRUE,
  ...
)
}
\description{
Specify the y-axis
}
